+++ PHPUnit/Extensions/Selenium2TestCase.php
@@ -54,6 +54,7 @@
  * @version    Release: @package_version@
  * @link       http://www.phpunit.de/
  * @since      Class available since Release 1.2.0
+ * @method void back()
  * @method \PHPUnit_Extensions_Selenium2TestCase_Element byCssSelector($value)
  * @method \PHPUnit_Extensions_Selenium2TestCase_Element byClassName($vaue)
  * @method \PHPUnit_Extensions_Selenium2TestCase_Element byId($value)