+++ ChangeLog.markdown
@@ -9,6 +9,7 @@ PHPUnit_Selenium 1.2.4
 * Supporting form submit (also via children elements).
 * Supporting radio boxes.
 * Supporting implicit waits.
+* Supporting back button.
 * Correctly marking Selenium 1 tests as skipped when server is not running.
 
 PHPUnit_Selenium 1.2.3